Added features to Game
Added use for size field, now used in logic for printing
Added changes to win condition. If exit is passed, player data becomes
null.
Added printGame to Game, this prints the game board in the formatting
used for simulation gameplay
Added containsExit method that returns true if the given cell contains
the exit
Added setAgentBeginning function to Game which sets the agent's
currentRow and currentCell to be the bottom right of the board